+ добавил метод LogAsync в класс сервиса для логирования непосредственно запроса, логику с мидлвара положил туда + описал делегат в конструкции use. такую конструкцию можно использовать как альтернативу
