+++ LessonMonitor/LessonMonitor.API/Services/ILogService.cs
@@ -1,7 +1,11 @@
-﻿namespace LessonMonitor.API.Services
+﻿using Microsoft.AspNetCore.Http;
+using System.Threading.Tasks;
+
+namespace LessonMonitor.API.Services
 {
     public interface ILogService
     {
         void Log(string message);
+        Task LogAsync(HttpRequest request);
     }
 }
